react router v6.8 (#4026)
* setup router compatibility package

* migrate routes

* migrate routes & cleanup

* migrate routes & cleanup

* migrate routes & cleanup

* react-router@6.3 everywhere. no route leaving guard.

* dev proxy for grafana

* react-router@6.3 everywhere. no route leaving guard. fixing bugs. refactor.

* upgrade rr to latest

* fix routes in test

* fix modal error

* road to data router

* RouteLeavingGuard fixed

* cleanup

* cleanup

* cleanup

* revert publicPath use

+++ designer/client/components/ProcessComments.tsx
@@ -42,8 +42,10 @@ function ProcessComments(): JSX.Element {
       text: DialogMessages.deleteComment(),
       confirmText: "DELETE",
       denyText: "NO",
-      onConfirmCallback: async () => {
-        await dispatch(deleteComment(processId, comment.id))
+      onConfirmCallback: async (confirmed) => {
+        if (confirmed) {
+          await dispatch(deleteComment(processId, comment.id))
+        }
         setPending(false)
       },
     })
